src/main/java/casting/DateToLocalDateTİme.java
package casting;

import java.time.LocalDateTime;
import java.time.ZoneId;
import java.util.Date;

public class DateToLocalDateTİme {
    public static void main(String[] args) {
        Date date = new Date();
        System.out.println(convertToLocalDateTimeViaInstant(date));
    }

    public static LocalDateTime convertToLocalDateTimeViaInstant(Date dateToConvert) {
        return dateToConvert.toInstant()
                .atZone(ZoneId.systemDefault())
                .toLocalDateTime();
    }
}
